move data source parsing
